requirements.txt
djangorestframework
djangorestframework-simplejwt
pillow
python-dotenv
pymysql
django-cors-headers
